Update awaySwitchAccessory.ts
I believe the idea to set "home" as an indefinite hold is flawed. This method means we'll only adhere to one heat/cool range all day, year round regardless of our sleep preferences. When we are home, we'll want the home, afternoon and sleep schedules to operate as we designed with in the Ecobee.

Having the "away" on indefinite makes sense and should stay as is.

This allows the behavior of "home" to resume schedule as an option in the configuration page for the plugin.